# internal/zikr/port/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Created by Otajonov Quvonchbek and Usmonov Azizbek",
        "contact": {}
    },
    "paths": {
        "/count": {
            "post": {
                "description": "This API create zikr count",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "count"
                ],
                "summary": "Create zikr count",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Count"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "This API updates zikr count",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "count"
                ],
                "summary": "Update zikr count",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Count"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/create-user": {
            "post": {
                "description": "register-user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get or Create user",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/model.UserGuid"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/zikr/create": {
            "post": {
                "description": "create-zikr",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "zikr"
                ],
                "summary": "Create zikr",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Zikr"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/zikr/get-all": {
            "get": {
                "description": "This API gets a list of zikr",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "zikr"
                ],
                "summary": "Get zikr list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GUID of the user",
                        "name": "guid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/model.Zikrs"
                        }
                    },
                    "404": {
                        "description": "Error response",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Count": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "user_guid": {
                    "type": "string"
                },
                "zikr_guid": {
                    "type": "string"
                }
            }
        },
        "model.GetZikr": {
            "type": "object",
            "properties": {
                "arabic": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "guid": {
                    "type": "string"
                },
                "is_favorite": {
                    "type": "boolean"
                },
                "pronounce": {
                    "type": "string"
                },
                "uzbek": {
                    "type": "string"
                }
            }
        },
        "model.Response": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        },
        "model.UserGuid": {
            "type": "object",
            "properties": {
                "guid": {
                    "type": "string"
                }
            }
        },
        "model.UserLoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.Zikr": {
            "type": "object",
            "properties": {
                "arabic": {
                    "type": "string"
                },
                "pronounce": {
                    "type": "string"
                },
                "uzbek": {
                    "type": "string"
                }
            }
        },
        "model.Zikrs": {
            "type": "object",
            "properties": {
                "zikrs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GetZikr"
                    }
                }
            }
        }
    }
}